+++ src/components/Courses.js
@@ -136,7 +136,7 @@ class Courses extends React.Component {
 
   renderIsEmptyMessage = (query, source) => {
     if (!source) {
-      return 'no source yet';
+      return <FormattedMessage id="ui-courses.noSource" />;
     }
 
     return (