+++ src/components/ReserveForm/sections/ReserveFormCopyright.js
@@ -8,11 +8,9 @@ import { Card, Col, Row, TextField, Select, Checkbox } from '@folio/stripes/comp
 
 const ReserveFormCopyright = (props) => {
   const { data } = props;
-  // const trackingObject = props.values.copyrightTracking || {};
-  // console.log('ReserveFormCopyright: props =', props);
 
   return (
-    <Card headerStart="Copyright information">
+    <Card headerStart={<FormattedMessage id="ui-courses.headings.copyrightInformation" />}>
       {/*
       <Row>
         <Col xs={12}>
@@ -20,7 +18,7 @@ const ReserveFormCopyright = (props) => {
             id="edit-reserve-copyright-tracking"
             name="copyrightTracking."
             component={TextField}
-            label="Copyright tracking required?"
+            label={<FormattedMessage id="ui-courses.field.copyrightTrackingNeeded" />}
           />
         </Col>
       </Row>