+++ src/components/ViewCourse/sections/index.js
@@ -3,4 +3,3 @@ export { default as ViewCourseCrosslistings } from './ViewCourseCrosslistings';
 export { default as ViewCourseInstructors } from './ViewCourseInstructors';
 export { default as ViewCourseTerm } from './ViewCourseTerm';
 export { default as ViewCourseReserves } from './ViewCourseReserves';
-export { default as ViewCourseDeveloper } from './ViewCourseDeveloper';